app/requirements.txt
aiohttp==3.7.4.post0
APScheduler==3.6.3
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.4.0
bleach==4.1.0
boto3==1.21.36
botocore==1.24.36
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
cryptography==36.0.1
defusedxml==0.7.1
discord.py==1.7.3
Django==3.2.6
django-allauth==0.47.0
django-appconf==1.0.5
django-axes==5.33.0
django-ckeditor==6.2.0
django-compat==1.0.15
django-crontab==0.7.1
django-ipware==4.0.2
django-js-asset==2.0.0
django-magiclink==1.2.0
django-recaptcha==3.0.0
django-social-share==2.2.1
django-storages==1.12.3
django-summernote==0.8.20.0
djangorestframework==3.13.1
fancycompleter==0.9.1
gunicorn==20.1.0
idna==3.3
jmespath==1.0.0
multidict==6.0.2
oauthlib==3.1.1
packaging==20.9
pdbpp==0.10.3
Pillow==9.0.0
psycopg2-binary==2.9.1
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
pyparsing==3.0.7
pyrepl==0.9.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-telegram-bot==13.11
python3-openid==3.2.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
rcssmin==1.1.0
requests==2.27.1
requests-oauthlib==1.3.0
rjsmin==1.2.0
s3transfer==0.5.2
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
stripe==2.65.0
tornado==6.1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.8
webencodings==0.5.1
wmctrl==0.4
yarl==1.7.2
